refactor(kilnd): extract a library target so integration tests can run (#377)
kilnd was a bin-only crate with its config/engine types defined inline in
main.rs (1618 LOC), so `tests/` integration files could not import them — the
two integration test files had been dark (uncompilable) since the WRT→Kiln
rename, giving zero coverage.

Split the crate:
- `src/lib.rs` — the library: KilndConfig/KilndEngine/WasiVersion/MemoryProfiler/
  RuntimeStats + the modules, and `pub fn run()` (the former `main_with_stack`).
- `src/main.rs` — a thin binary owning only the process-level items a binary must
  define: the no_std `#[global_allocator]`, the `#[panic_handler]`, and the
  `fn main` entry points, delegating to `kilnd::run()`.
- Cargo.toml gains a `[lib]` target.

Restored `integrated_features_tests.rs` to the current API (imports via `kilnd::`,
`WasiCapabilities::minimal()?`, `Preview1`→`Preview2`, `..Default::default()`,
`'static` fixtures) and to current *correct* behavior (fail-loud on an
entry-point-less module; assert registration happened rather than a brittle
count; error on a full-width non-WASM binary). Marked the tests
`#[serial_test::serial]` — they construct KilndEngine, which draws on the global
capability memory budget and must not contend in parallel. Result: 13/13 pass.

Deleted `kilnd_integration_tests.rs`: it tests a removed API — `KilnDaemon`
(gone), `daemon.run(config)` (no such method), and a KilndConfig with seven
fields that no longer exist. It can't be refactor-fixed (nothing to import) and
rewriting it would be authoring new tests; per CLAUDE.md a broken dark test is
worse than none.

`detect_component_format` made pub (the engine method the tests exercise). The
pre-existing `--no-default-features` (no_std) build breakage is unchanged —
identical `self.memory_profiler` field/method inconsistency exists on main; it is
not in CI and out of scope here.

Trace: SR-27
